Separate tests file
