+++ polls/tests/test_index.py
@@ -1,0 +1,73 @@
+import datetime
+
+from django.test import TestCase
+from django.urls import reverse
+from django.utils import timezone
+
+from polls.models import Question
+
+
+def create_question(question_text, days):
+    """Create a question.
+
+    with the given `question_text` and published the
+    given number of `days` offset to now (negative for questions published
+    in the past, positive for questions that have yet to be published).
+
+    Returns:
+        Question : a new question
+    """
+    pub = timezone.now() + datetime.timedelta(days=days)
+    end = pub+datetime.timedelta(days=365)
+    return Question.objects.create(
+        question_text=question_text,
+        pub_date=pub,
+        end_date=end
+    )
+
+
+class IndexTests(TestCase):
+    """Tests of index view."""
+
+    def test_no_questions(self):
+        """Test if no questions exist, an appropriate message is displayed."""
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """Test for questions with a pub_date in the past are displayed on the index page."""
+        create_question(question_text="Past question.", days=-30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_future_question(self):
+        """Test for questions with a pub_date in the future aren't displayed on the index page."""
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """Test even if both past and future questions exist, only past questions are displayed."""
+        create_question(question_text="Past question.", days=-30)
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_two_past_questions(self):
+        """Test for the questions index page may display multiple questions."""
+        create_question(question_text="Past question 1.", days=-30)
+        create_question(question_text="Past question 2.", days=-5)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question 2.>', '<Question: Past question 1.>']
+        )
