+++ polls/tests/test_poll_dates.py
@@ -1,14 +1,12 @@
-"""Tests for polls app."""
 import datetime
 
 from django.test import TestCase
 from django.utils import timezone
-from django.urls import reverse
 
-from .models import Question
+from polls.models import Question
 
 
-class QuestionModelTests(TestCase):
+class PollDatesTests(TestCase):
     """Tests of question model."""
 
     def setUp(self):
@@ -101,94 +99,3 @@ class QuestionModelTests(TestCase):
         self.assertIs(question5.can_vote(), False)
         question6 = Question(pub_date=day_before, end_date=time)
         self.assertIs(question6.can_vote(), False)
-
-
-def create_question(question_text, days):
-    """Create a question.
-
-    with the given `question_text` and published the
-    given number of `days` offset to now (negative for questions published
-    in the past, positive for questions that have yet to be published).
-
-    Returns:
-        Question : a new question
-    """
-    pub = timezone.now() + datetime.timedelta(days=days)
-    end = pub+datetime.timedelta(days=365)
-    return Question.objects.create(
-        question_text=question_text,
-        pub_date=pub,
-        end_date=end
-    )
-
-
-class QuestionIndexViewTests(TestCase):
-    """Tests of index view."""
-
-    def test_no_questions(self):
-        """Test if no questions exist, an appropriate message is displayed."""
-        response = self.client.get(reverse('polls:index'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_past_question(self):
-        """Test for questions with a pub_date in the past are displayed on the index page."""
-        create_question(question_text="Past question.", days=-30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_future_question(self):
-        """Test for questions with a pub_date in the future aren't displayed on the index page."""
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_future_question_and_past_question(self):
-        """Test even if both past and future questions exist, only past questions are displayed."""
-        create_question(question_text="Past question.", days=-30)
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_two_past_questions(self):
-        """Test for the questions index page may display multiple questions."""
-        create_question(question_text="Past question 1.", days=-30)
-        create_question(question_text="Past question 2.", days=-5)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question 2.>', '<Question: Past question 1.>']
-        )
-
-
-class QuestionDetailViewTests(TestCase):
-    """Tests of detail view."""
-
-    def test_future_question(self):
-        """Test for the detail view of a question with a pub_date in the future.
-
-        It must returns a 404 not found.
-        """
-        text = 'Future question'
-        future_question = create_question(question_text=text, days=5)
-        url = reverse('polls:detail', args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-
-    def test_past_question(self):
-        """Test for the detail view of a question with a pub_date in the past.
-
-        It must displays the question's text.
-        """
-        past_question = create_question(question_text='Past Question', days=-5)
-        url = reverse('polls:detail', args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertContains(response, past_question.question_text)
